Rename to generic AgentOptimizer
